set flamby models separately from others
