+++ examples/flamby_mpi_sync.py
@@ -7,7 +7,7 @@ from mpi4py import MPI
 from appfl.config import *
 from appfl.misc.data import *
 from appfl.misc.utils import *
-from models.utils import flamby_train
+from models.flamby import flamby_train
 from dataloader.flamby_dataloader import get_flamby
 
 """
